Move general renaming functions to prover package
