+++ src/main/scala/at/logic/skeptik/algorithm/prover/CR.scala
@@ -2,9 +2,8 @@ package at.logic.skeptik.algorithm.prover
 
 import at.logic.skeptik.algorithm.prover.structure.immutable.Literal
 import at.logic.skeptik.algorithm.unifier.{MartelliMontanari => unify}
-import at.logic.skeptik.expression.{Abs, App, E, Var, i}
 import at.logic.skeptik.expression.substitution.immutable.Substitution
-import at.logic.skeptik.expression.substitution.mutable.{Substitution => MSubstitution}
+import at.logic.skeptik.expression.{E, Var}
 
 import scala.collection.mutable
 import scala.collection.mutable.ArrayBuffer
@@ -146,55 +145,6 @@ object CR {
           }
         }
       }
-    }
-
-    /**
-      * Gets all unifiable variables (contained in `variables`) from
-      * given Es.
-      *
-      * @param exps where unifiable variables should be found
-      * @return unifiable variables contained at least once in exps
-      */
-    def unifiableVars(exps: E*): Set[Var] = exps.flatMap {
-      case App(e1, e2) =>
-        unifiableVars(e1) union unifiableVars(e2)
-      case Abs(v, e1) =>
-        unifiableVars(v) union unifiableVars(e1)
-      case v: Var =>
-        if (variables contains v) Set(v) else Set.empty[Var]
-    }.toSet
-
-    /**
-      * Rename quantified variables in left so that they don't intersect
-      * with quantified variables in right. It's necessary for unification
-      * to work correctly.
-      *
-      * @param left where quantified variables should be renamed
-      * @param usedVars already used variables
-      * @return proper substitution to rename without variable collisions
-      */
-    def renameVars(left: E, usedVars: Set[Var]): Substitution = {
-      val sharedVars = unifiableVars(left) intersect usedVars // Variables which should be renamed
-
-      // Unification variables which can be reused for new variables
-      val notUsedVars = variables diff (sharedVars union unifiableVars(left) union usedVars)
-
-      val kvs = for (v <- sharedVars) yield {
-        val replacement = notUsedVars.headOption getOrElse { // Use some variable from unification variables
-          // Or create a new one
-          var newVar = Var(v + "'", i)
-          while (sharedVars contains newVar) {
-            newVar = Var(newVar + "'", i)
-          }
-          variables += newVar // It will be available for unification from now
-          newVar
-        }
-
-        if (notUsedVars contains replacement) notUsedVars -= replacement
-        v -> replacement
-      }
-
-      new Substitution(kvs.toMap)
     }
 
     /**
